fix warnings

+++ src/winapi_hooks.c
@@ -807,8 +807,8 @@ BOOL WINAPI fake_StretchBlt(
         {
             return real_StretchBlt(
                 g_ddraw.render.hdc,
-                (xDest * g_ddraw.render.scale_w) + g_ddraw.render.viewport.x,
-                (yDest * g_ddraw.render.scale_w) + g_ddraw.render.viewport.y,
+                (int)(xDest * g_ddraw.render.scale_w) + g_ddraw.render.viewport.x,
+                (int)(yDest * g_ddraw.render.scale_w) + g_ddraw.render.viewport.y,
                 (int)(wDest * g_ddraw.render.scale_w),
                 (int)(hDest * g_ddraw.render.scale_h),
                 hdcSrc,
@@ -883,8 +883,8 @@ BOOL WINAPI fake_BitBlt(
         {
             return real_StretchBlt(
                 g_ddraw.render.hdc,
-                (x * g_ddraw.render.scale_w) + g_ddraw.render.viewport.x,
-                (y * g_ddraw.render.scale_w) + g_ddraw.render.viewport.y,
+                (int)(x * g_ddraw.render.scale_w) + g_ddraw.render.viewport.x,
+                (int)(y * g_ddraw.render.scale_w) + g_ddraw.render.viewport.y,
                 (int)(cx * g_ddraw.render.scale_w),
                 (int)(cy * g_ddraw.render.scale_h),
                 hdcSrc,
@@ -1026,8 +1026,8 @@ int WINAPI fake_StretchDIBits(
             return
                 real_StretchDIBits(
                     g_ddraw.render.hdc,
-                    (xDest * g_ddraw.render.scale_w) + g_ddraw.render.viewport.x,
-                    (yDest * g_ddraw.render.scale_h) + g_ddraw.render.viewport.y,
+                    (int)(xDest * g_ddraw.render.scale_w) + g_ddraw.render.viewport.x,
+                    (int)(yDest * g_ddraw.render.scale_h) + g_ddraw.render.viewport.y,
                     (int)(DestWidth * g_ddraw.render.scale_w),
                     (int)(DestHeight * g_ddraw.render.scale_h),
                     xSrc,
